🐞 [Anomaly Task] Fix progress bar (#1223)
* Address progress issue for HPO callback

* Merge hpo and progress callback as only one progress callback is supported
