+++ external/anomaly/tasks/train.py
@@ -16,7 +16,7 @@
 
 from typing import Optional
 
-from adapters.anomalib.callbacks import ProgressCallback, ScoreReportingCallback
+from adapters.anomalib.callbacks import ProgressCallback
 from adapters.anomalib.data import OTEAnomalyDataModule
 from adapters.anomalib.logger import get_logger
 from anomalib.utils.callbacks import (
@@ -68,7 +68,6 @@ class TrainingTask(InferenceTask, ITrainingTask):
         callbacks = [
             ProgressCallback(parameters=train_parameters),
             MinMaxNormalizationCallback(),
-            ScoreReportingCallback(parameters=train_parameters),
             MetricsConfigurationCallback(
                 adaptive_threshold=config.metrics.threshold.adaptive,
                 default_image_threshold=config.metrics.threshold.image_default,
